Add a simple test that calls dtrace_probe().
Currently, we have no way of knowing whether or not this actually
works because we're not checking any output. This only indicates
that the code does not crash -- not that it does the correct thing.

Further work needs to be done in separating the DIF interpreter from
dtrace_probe() entirely and trying to re-write it in Haskell so that
we can run some fuzz tests on it.