Some of the fuzzes were useless.
We both register and unregister in the dtrace_probe_create fuzz, so
there is no need to do it separately at this point. Move over the
code from dtrace_probe_create to init which stands for initialization
and fuzz that instead. It uses less CPU resources all together and
doesn't require multiple fuzzes to be running at the same time, which
may result in duplicate crashes.