Initial implementation of FileCheck-based tests.
This is something that will be used to test the D compiler and currently
requires root. In order to fix this, we should separate out DTrace into
an engine and create a process that will contain the probes and talk
to the testing binary.

In addition to that, the way these tests are run is subject to change.